Change unsigned long to size_t

+++ src/mod_manager.h
@@ -169,8 +169,8 @@ class mod_ui
 
         void try_add( const std::string &mod_to_add,
                       std::vector<std::string> &active_list );
-        void try_rem( unsigned long selection, std::vector<std::string> &active_list );
-        void try_shift( char direction, unsigned long &selection, std::vector<std::string> &active_list );
+        void try_rem( size_t selection, std::vector<std::string> &active_list );
+        void try_shift( char direction, size_t &selection, std::vector<std::string> &active_list );
 
         bool can_shift_up( long selection, std::vector<std::string> active_list );
         bool can_shift_down( long selection, std::vector<std::string> active_list );
